Remove dangling code in sdk/
