+++ hack/load.cli.ts
@@ -7,9 +7,9 @@ import * as os from "node:os";
 import * as path from "node:path";
 import * as fs from "node:fs/promises";
 import * as R from "ramda";
-import { heredoc } from "../src/sdk/heredoc";
 import * as lib from "./load.lib";
 import { execSync } from "node:child_process";
+import { heredoc } from "./load.lib";
 const TEST_CLUSTER_NAME_PREFIX = "pepr-load";
 const TEST_CLUSTER_NAME_DEFAULT = "cluster";
 const TEST_CLUSTER_NAME_MAX_LENGTH = 32;
